lmao/requirements.txt
Django==2.2.7
djangorestframework==3.10.3
Markdown==3.1.1
numpy==1.17.4
Pillow==6.2.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
torch==1.3.1+cpu
torchvision==0.4.2+cpu
